move world-map to it's own component
