+++ test/components/world.spec.js
@@ -7,7 +7,7 @@ import World from '../../app/scripts/components/world'
 describe('World', () => {
   it('should have the correct defaults', () => {
     const el = shallow(<World/>)
-    expect(el.find('AutoSizer').length).to.equal(1)
+    expect(el.find('WorldMap').length).to.equal(1)
 
     var inst = el.instance()
     expect(inst.props.peersCount).to.equal(0)
